au_bot/requirements.txt
certifi==2020.11.8
chardet==3.0.4
idna==2.10
install==1.3.4
oauthlib==3.1.0
pip==20.2.4
PySocks==1.7.1
requests==2.25.0
requests-oauthlib==1.3.0
setuptools==47.1.0
six==1.15.0
tweepy==3.9.0
urllib3==1.26.2
python-dateutil==2.8.1
python-decouple==3.3
